Refactor model imports to use generation submodule
