+++ models/__init__.py
@@ -1,6 +1,6 @@
-from models.api_model import APIModel
-from models.openai_model import OpenAIModel
-from models.model_factory import ModelFactory
+from models.generation.api_model import APIModel
+from models.generation.openai_model import OpenAIModel
+from models.generation.model_factory import ModelFactory
 
 __all__ = [
     "APIModel",
